Everything is a Guage
Because we are sending stats to StatsD on an interval everything is a gauge
even "counts" which are ever increasing. There's a current issue open for
StatsD (etsy/statsd#324) which requests "absolute" counters which would
solve this problem however until that's solved we have to use gauges and
just apply derivative functions after the fact. Le Sigh.